test: create unit test for sort service
